authen-api

+++ app/Http/Controllers/API/ArtistController.php
@@ -6,6 +6,7 @@ use App\Http\Controllers\Controller;
 use App\Http\Resources\ArtistResource;
 use App\Models\Artist;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 use Illuminate\Validation\Rule;
 
 class ArtistController extends Controller
@@ -15,6 +16,7 @@ class ArtistController extends Controller
      */
     public function index()
     {
+        Gate::authorize('viewAny', Artist::class);
         $artists = Artist::query()->paginate(10);
         return ArtistResource::collection($artists);
     }
@@ -24,6 +26,7 @@ class ArtistController extends Controller
      */
     public function store(Request $request)
     {
+        Gate::authorize('create', Artist::class);
         $request->validate([
             'name' => ['required', 'max:255', 'unique:artists,name'],
             'image_path' => ['nullable', 'image', 'mimes:jpeg,png,jpg,gif,svg'],
@@ -44,6 +47,7 @@ class ArtistController extends Controller
      */
     public function show(Artist $artist)
     {
+        Gate::authorize('view', $artist);
         return new ArtistResource($artist);
     }
 
@@ -52,6 +56,7 @@ class ArtistController extends Controller
      */
     public function update(Request $request, Artist $artist)
     {
+        Gate::authorize('update', $artist);
         $request->validate([
             'name' => [
                 'required', 'max:255',
@@ -75,6 +80,7 @@ class ArtistController extends Controller
      */
     public function destroy(Artist $artist)
     {
+        Gate::authorize('delete', $artist);
         $artist->delete();
         return response(null, 204);
     }
